Initial Commit of LKit
This is the initial commit for LKit, a simple lisp-like language that I
wanted to create to make embedding a parser into projects as simple as
possible. The data is primarily just int and doubles, but I've left in the
timestamp as well to make timeseries data sets something to expand into.