Added clone() Method to Functions
Given the memory management scheme we're going to have on the parser, it made
a lot of sense to have a clone() method on the functions as well as the
values so that we can make deep-copies of the parser and it's contents. This
is just a more convenient scheme than trying to know if it's a reference or
what. Everything is a pointer, and all objects have clone(). Simple.