Added More Functions to Base Set
I wanted to be able to have a decent initial set, so I added in the regular
arithmetic set (+, -, *, /) and then added in the comparison operators as
well as the boolean checks (and, or, not) so that we have a pretty decent
initial set to work with. In order to save code, the comparison and binary
tests are really just one class, but an ivar tells it how to process the
data for the type of test it's doing. This makes the code slightly more
complex, but not too bad, and the savings in lines of code are significant.