Added Evaluation of Values in Functions
I was missing a key component in the base functions - each argument needed
to be evaluated BEFORE it was tested so that if it happened to be an
expression, it was evaluated before it was used. This isn't a short-cut on
what needs to be calculated - not yet, but it's an important step in making
sure that all sub-trees are being calculated properly.