Added a toString() Method that didn't Lock
I needed to have a "no-lock" version of toString() for a subclass, and so I
just made it like the other no-lock versions. This doesn't really have any
significant overhead, and it makes it possible to get the toString() value
within a locked section of code. Very important at times, it turns out.