Added eval() and Boolean Operators
I needed to have the ability to eval() and get a value instance back (as a
copy). So I added that as a locking and no-lock versions. Simple, but it is
going to prove very useful in the expression evaluation in the functions.
Secondly, I added in the boolean operators !, && and || that return standard
bool values so that the logical construction of these values works as you'd
expect. Finally, I had to add the const-ness to the equality/inequality
operators as there was an ambiguity in the compiler as to which one should
be used, and this clears all that up.