Added the Named Constructor
I needed to be able to create a variable with the proper name but without
a value - so I created a new constructor. This makes the parser able to
detect undefined variables and create them with no value, so that the caller
can find out what he needs to provide, and provide it.