Added Tests for Timer's Parsing
I wanted to have tests for the new parsing functions of the timer class, so
I made them. There's a bit of a question about what to use as the date when
you are only given a time, but I'm leaving it at epoch as that's the most
understandable reference point for such a thing.