Created Parser Tests
I needed to add a few tests for the parser so that I could be sure I was
getting the right answers. Turns out, I needed to fix a few things, but that's
what the tests are for, no? In the end, I'm sure things are processing properly
and the expressions are being evaluated properly.