Created a Parser for LKit
I needed to have something that would take a std::string that could come from
about anywhere, and create a language tree of expressions, variables, functions
and the like, that can then be evaluated and the value returned to the caller.
This wasn't too bad, as the language is meant to be pretty simple. I just look
at what's in the stream and go from there.