Updated the README's TODO List
Now that I have the first cut of the parser done, I needed to update the things
I wanted to do next. Certainly the variables and functions need to be able to
be defined, and that's next, but after that, there's still a lot of neat and
fun things to do.