Added Value Pointer Support to Variable
When we wanted to have an expression held in a variable we were in a bit of a
jam. There was no way to have an expression variable - until now. What I did
was to put into the variable a value pointer and then brace up the methods
so that we could set an expression as a variable's "value". This was really
necessary as we needed to have this capability.