Added Classification Methods and No-Lock toString
I needed to be able to have the equivalent of Java's instanceof method, and
by far the fastest and easiest was to have a series of isA() methods that
are all set to return 'false' except for the one that the subclass is. So I
added isVariable() and isExpression() and then implemented them to return
'true' in those respective subclasses of value. At the same time, I needed to
have a no-lock version of toString() as the subclasses needed to be able to
call toString() without it locking. This has been in a few classes, and I just
put it in here as well.