Added in a New Constructor for Variable
I didn't want to have to make a pointer for a simple variable creation, so
I made a new version of the variable constructor that takes a reference to a
value and that will allow the compiler to take anything that can be made INTO
a value and make it a value and then pass it into the constructor for the
value. This just makes the construction of variables a little cleaner.