Added Multiple Top-Level Expressions to Parser
I wanted to be able to have multiple top-level expressions so that I could have
multiple expressions in the parser and have them all evaluated one after
another. In the end, the final value is all I want, but this sets the stage
for the creation of functions like lisp, and then the use of those functions.
This is just the step I needed to take.